requirements.txt
faster_whisper==1.0.3
langchain_groq==0.2.0
numpy==2.1.2
openai==1.52.0
playsound==1.3.0
pynput==1.7.7
python-dotenv==1.0.1
scipy==1.14.1
sounddevice==0.4.7
gtts==2.5.3
pygame==2.6.1